refactor feature flag
